Add files via upload
Console commands in main are now separate functions. Reversed some "compressions" of loops and conditionals, where they were reduced to one-row code segments which could be considered hard to read. Changed the method of field generation, where instead of checking if a given tile is in a corner, now a new table is created, which is padded with an extra row/column of zeroes on each side and those checks are now obsolete, hence shedding some extra code.